refactored db.go into 6 files for reading data from firebase relating to courses, schools, professors, grades, apikeys, and general db functions
